requirements.txt
beautifulsoup4
requests
datetime
flask
Flask-HTTPAuth
gunicorn
mysql-connector-python